new iOS schema

+++ app/workers/cocoapod_service_worker.rb
@@ -16,11 +16,11 @@ class CocoapodServiceWorker
 
     ActiveRecord::Base.logger.level = 1
 
-    cocoapods_url = "https://search.cocoapods.org/api/v1/pods.picky.hash.json"
+    base = "https://search.cocoapods.org/api/v1/pods.picky.hash.json"
 
-    url = "#{cocoapods_url}?query=on%3Aios+#{char}&ids=#{res_count}&offset=#{offset}&sort=name"
+    url = "#{base}?query=on%3Aios+#{char}&ids=#{res_count}&offset=#{offset}&sort=name"
 
-    puts "Scraping #{char}, #{res_count} - #{res_count.to_int + offset.to_int}"
+    puts "Scraping #{char}, #{offset} - #{res_count.to_int + offset.to_int}"
     
     begin
 
@@ -38,40 +38,46 @@ class CocoapodServiceWorker
 
       5.times do |i|
 
-        if results[i].present?
-        
-          res = results[i][5]
+        next if results[i].nil?
+      
+        res = results[i][5]
 
-          if res.present?
+        next if res.nil?
 
-            res.each do |result|
+        res.each do |r|
 
-              begin
+          begin
 
-                cp = Cocoapod.create(git: result['source']['git'], http: result['source']['http'], tag: result['source']['tag'], summary: result['summary'], link: result['link'], cocoadocs: result['cocoadocs'], name: result['id'], version: result['version'])
+            s = r['source']
+            git = s['git']
+            http = s['http']
+            tag = s['tag']
+            summ = r['summary']
+            link = r['link']
+            cdocs = r['cocoadocs']
+            name = r['id']
+            ver = r['version']
 
-                result['authors'].each do |name, email|
+            cp = Cocoapod.create(git: git, http: http, tag: tag, summary: summ, link: link, cocoadocs: cdocs, name: name, version: ver)
 
-                  CocoapodAuthor.find_or_create_by(name: name, email: email, cocoapod: cp)
-                
-                end
+            r['authors'].each do |name, email|
 
-                result['tags'].each do |tag|
-
-                  CocoapodTag.create(tag: tag, cocoapod: cp)
-
-                end
-
-              rescue
-
-                nil
-
-              end
-              
+              CocoapodAuthor.find_or_create_by(name: name, email: email, cocoapod: cp)
+            
             end
 
-          end
+            r['tags'].each do |tag|
 
+              CocoapodTag.create(tag: tag, cocoapod: cp)
+
+            end
+
+          rescue
+
+            nil
+
+          end
+          
         end
 
       end
@@ -108,13 +114,11 @@ class CocoapodServiceWorker
 
     return nil if cocoapod.nil? || source_code_url.nil?
 
-    ext = '.zip'
-
     basename = cocoapod.name
 
-    dump = Rails.env.production? ? '"/mnt/sdk_dump/"' : '../sdk_dump/'
+    dump = Rails.env.production? ? '/mnt/sdk_dump/' : '../sdk_dump/'
 
-    filename = dump + basename + ext
+    filename = dump + basename + '.zip'
 
     headers = {
       'Content-Type' => 'application/zip',
@@ -153,11 +157,15 @@ class CocoapodServiceWorker
 
   def parse_header(filename:, cocoapod_id:)
 
+    return nil unless File.exist? filename
+
     file = File.open(filename).read
 
     names = file.scan(/(@interface|@protocol)\s(.*?)[^a-zA-Z]/i).uniq
 
     names.each do |name|
+
+      next if in_apple_docs?(name[1]) || name[1].blank?
 
       begin
 
@@ -168,6 +176,34 @@ class CocoapodServiceWorker
         nil
 
       end
+
+    end
+
+  end
+
+  def in_apple_docs?(q)
+
+    apple_docs = AppleDoc.find_by_name(q)
+
+    return true if apple_docs.present?
+
+    uri = URI("https://developer.apple.com/search/search_data.php")
+
+    headers = { 'User-Agent' => UserAgent.random_web }
+
+    data = Proxy.get(req: {:host => uri.host, :path => uri.path, :protocol => uri.scheme, :headers => headers}, params: {'q' => q})
+
+    in_docs = JSON.parse(data.body).any?{ |res| res['description'].downcase.include? q.downcase }
+
+    if in_docs
+
+      AppleDoc.create(name: q)
+
+      true
+
+    else
+
+      false
 
     end
